groq-builder: feature(compat): moved all exports to the main index, removing the need for the `/validation` entrypoint
