--- packages/groq-builder/validation/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "../dist/validation/index"
-}
